add indicator longdesc, refactor main into modules
