+++ scripts/mkess/dataset/zapspecial.go
@@ -1,4 +1,4 @@
-package main
+package dataset
 
 import "strings"
